Clean up S3 Buckets in integration test (#1603)
### Feature or Bugfix
- Feature

### Detail
Environment and Dataset stacks leave S3 Buckets to be deleted even when
the CloudFormation stack is deleted.
This PR deletes the S3 Buckets when the session and temp fixtures for
datasets and environments are deleted.

### Testing
Tested that environment and dataset buckets are all deleted - in real
AWS CICD pipeline with NOT-empty buckets.
Tested that dataset buckets with some manually created access points
succeed.

### Relates
- #1220 

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.