Integration tests executed on a real deployment as part of the CICD - Redshift Connections (#1628)
⚠️ NEEDS SOME CHANGES AFTER #1638 is merged

### Feature or Bugfix
- Feature: Testing

### Detail
Add integration tests for Redshift connections
It also includes the global conftest fixtures that will be used for
redshift-datasets and redshift dataset sharing

At the moment it assumes Redshift infrastructure is already provided; we
might want to implement the deployment of this pre-required infra, but
that can be done on a separate PR.

+++ Additional fixes found during testing and fixed in this PR
- Fixed Pivot role permissions in CDK when no workgroups are defined (it
handled None wrong)
- Add update environment stack on Connection delete to ensure
permissions to pivot role are removed

### Testing
- Fixes deployed in CICD pipeline in AWS
- Tests executed locally (in this case we can be sure it will work as in
AWS as there are no AWS Clients, just API calls)



![image](https://github.com/user-attachments/assets/4968ec38-48c1-4314-b9b3-32992c692eb6)


### Relates
- #1619 
- #1220 

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.