User modal dialog - team link (#1627)
### Feature or Bugfix
- Feature

### Detail
- Turns various appearances of Team name text into clickable links,
which pop up a modal dialog that lists the users in the team.
- For more detail on the proposal, see:
https://github.com/data-dot-all/dataall/issues/1626
- As we don't have Cognito available to us and use a separate custom
system for roles management, we could use some help with feedback and
testing on the Cognito piece of this, if there is interest in accepting
this contribution.

### Relates
- https://github.com/data-dot-all/dataall/issues/1626

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)? NO
  - Is the input sanitized?  N/A
- What precautions are you taking before deserializing the data you
consume? N/A
  - Is injection prevented by parametrizing queries?  N/A
  - Have you ensured no `eval` or similar functions are used?  YES
- Does this PR introduce any functionality or component that requires
authorization? NO
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
N/A
  - Are you logging failed auth attempts?  N/A
- Are you using or adding any cryptographic features?  NO
  - Do you use a standard proven implementations?  N/A
- Are the used keys controlled by the customer? Where are they stored?
N/A
- Are you introducing any new policies/roles/users?  NO
  - Have you used the least-privilege principle? How?  N/A


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Co-authored-by: rbernota <rbernota@yahooinc.com>