Changes to the logic of calculating expiration date (#1635)
### Feature or Bugfix
- Feature 

### Detail

Changes to the logic of calculating expiration. 

Initially current month was considered into the calculation for
expiration date. But this resulted in a bug while performing unit tests
here https://github.com/data-dot-all/dataall/pull/1594 and also this was
not correct in the sense that if a user requests a month of expiration
at the end of month, that user will get expiration at the end of the
same month

But with the new logic -
https://github.com/data-dot-all/dataall/pull/1594, more than half a
month is granted extra .

**To minimize this extra access period on share, this PR proposes new
logic,**

1. If the user is in the last week of the month and if the user requests
1 month of data, allocate the end of the next month.
2. If the user is in NOT in the last week of the months and if the user
requests 1 month of data, allocate the end of same month as expiration.

### Relates
- https://github.com/data-dot-all/dataall/issues/1083

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Co-authored-by: trajopadhye <tejas.rajopadhye@yahooinc.com>