Fix: check if Redshift table exists before publishing it to data.all (#1644)
### Feature or Bugfix
- Bugfix

### Detail
Currently when we hit the addRedshiftTable api, we do not check if the
provided table exists in Redshift. In the UI the users must select from
a list that retrieves the Redshift tables, but programmatically they
could introduce any table, which can result in non-existent tables being
added to the data.all Catalog.

This PR 
1) checks if a table exists in Redshift and is accessible by the
Redshift connection before adding it to data.all
2) in the UI, it shows a success banner with the list of tables added
and an error banner with the list of error tables if any.
 
### Testing
- Tested in AWS 
- UI views show banner with added tables for importDataset and
addDataset
      - UI import Dataset does not show anything if no tables are added
- Added unit tests - checked as part of PR checks 🟢 
- Added integration tests - attached screenshot of redshift dataset
tests


![image](https://github.com/user-attachments/assets/e7e40b97-74ac-4d4c-9590-11c22055f524)

### Relates
- #1424 

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.