Added error view and unified utility to check tenant user (#1657)
### Feature or Bugfix
- Feature

### Detail
Users that do not belong to the DAAdministrators team can still navigate
to the admin console. They cannot do anything as all the API calls will
return errors. This PR takes it a step further and shows an error
message and does not render any of the components if the admin view.
Here is an snapshot of what users view:

I considered restricting the route in routes, but because we need to use
the useGroups hook, it was giving a lot of errors.



![image](https://github.com/user-attachments/assets/561dcb84-941b-4f01-8a62-18213364f5e8)


In addition, we have created a utility that unifies the checks to this
Cognito group in other places of the frontend.


### Relates


### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.