add custom domain support for apigw
Be aware that by enabling this feature the automatically generated apigw URL will be disabled.
As a prereq users must already have a custom domain in Route53 and have an ACM certificate.
The domain must be setup according to this https://docs.aws.amazon.com/apigateway/latest/developerguide/how-to-edge-optimized-custom-domain-name.html.
To enable this feature one must specify the following in cdk.json.

```
  "context": {
    "DeploymentEnvironments": [
      {
        ...
        "apigw_custom_domain": {
          "hosted_zone_name": "custom.domain.com",
          "hosted_zone_id": "..."
          "certificate_arn": "arn:aws:acm:us-east-1:..." # edge optimized domain thus cert must be in us-east-1
        }
      ...
    }
  }
```