Updating overly permissive policies tagged by checkov for environment role using least privilege principles (#1632)
### Feature or Bugfix
- Bugfix

### Detail
This PR is to restrict the overly permissive permissions in the
environment role IAM policies that are tagged as FAILURES by checkov.
Instead of using Get* and List* we specify the actions and type of
resources granted.

In addition, this PR uses KMS keys to encrypt the custom resources
Lambdas in the S3 datasets module

### Relates
- https://github.com/data-dot-all/dataall/issues/1524

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized? N/A
- What precautions are you taking before deserializing the data you
consume? N/A
  - Is injection prevented by parametrizing queries? N/A
  - Have you ensured no `eval` or similar functions are used? N/A
- Does this PR introduce any functionality or component that requires
authorization? N/A
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
N/A
  - Are you logging failed auth attempts? N/A
- Are you using or adding any cryptographic features? N/A
  - Do you use a standard proven implementations? N/A
- Are the used keys controlled by the customer? Where are they stored?
N/A
- Are you introducing any new policies/roles/users? N/A
- Have you used the least-privilege principle? How? Restricting overly
permissive permissions to the required resources / principle only.


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Co-authored-by: Mourya Darivemula <mouryacd@amazon.com>