tenant-permission tests  (#1694)
### Feature or Bugfix
- Feature

### Detail
Add unit tests that verify that MANAGE_X permissions are applied to all
Mutations except for an OPT_OUT list of Mutations and to a subset of
OPT_IN queries.

The OPT_OUT mutations are either:
- admin actions that can only be performed by the tenants. Applying
permissions in this case does not make sense.
- platform "support" features such as feed, notification, votes. No
object needs to be protected in this case.
- v2.7.0 features which will be addressed in a separate PR

The OPT_IN queries are operations that retrieve credentials or redirect
URLs that allow the user to effectively create/update data.all objects.

### Relates

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.